pyproject.toml
[project]
name = "autogen-exploration"
version = "0.1.0"
description = "Exloring the capabilities of AutoGen"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Indra Basak", email = "indra.basak1@gmail.com"}
]
dependencies = [
    "autogen-agentchat>=0.7.4",
    "autogen-ext[azure,langchain,magentic-one,mcp,openai]>=0.7.4",
    "langchain-community>=0.3.30",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.34",
    "pandas>=2.3.2",
    "pylint>=3.3.8",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.12",
    "sqlalchemy>=2.0.43",
]
